@@ -2,36 +2,36 @@
   "name": "v1ack.github.com",
   "version": "0.0.1",
   "scripts": {
-    "dev": "svelte-kit dev",
-    "build": "svelte-kit build",
-    "preview": "svelte-kit preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-    "format": "prettier --write --plugin-search-dir=. ."
+    "dev": "vite dev",
+    "build": "vite build",
+    "preview": "vite preview",
+    "prepare": "svelte-kit sync || echo ''",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+    "format": "prettier --write .",
+    "lint": "prettier --check ."
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "1.0.0-next.34",
-    "@sveltejs/kit": "1.0.0-next.350",
-    "@typescript-eslint/eslint-plugin": "^5.28.0",
-    "@typescript-eslint/parser": "^5.28.0",
-    "autoprefixer": "^10.4.7",
-    "postcss": "^8.3.1",
-    "postcss-import": "^14.1.0",
-    "postcss-load-config": "^4.0.1",
-    "postcss-scss": "^4.0.4",
-    "prettier": "~2.7.1",
-    "prettier-plugin-svelte": "^2.7.0",
-    "sass": "^1.52.3",
-    "svelte": "^3.49.0",
-    "svelte-check": "^2.7.2",
-    "svelte-preprocess": "^4.10.7",
-    "tailwindcss": "^3.1.3",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.4"
+    "@eslint/compat": "^1.2.5",
+    "@eslint/js": "^9.18.0",
+    "@fontsource/fira-mono": "^5.2.5",
+    "@sveltejs/adapter-static": "3.0.8",
+    "@sveltejs/kit": "^2.20.2",
+    "@sveltejs/vite-plugin-svelte": "^5.0.0",
+    "@tailwindcss/vite": "^4.0.0",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-svelte": "^3.0.0",
+    "globals": "^16.0.0",
+    "prettier": "~3.5.3",
+    "prettier-plugin-svelte": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.11",
+    "svelte": "^5.25.5",
+    "svelte-check": "^4.1.5",
+    "tailwindcss": "^4.0.17",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.20.0",
+    "vite": "^6.0.0"
   },
-  "type": "module",
-  "dependencies": {
-    "@fontsource/fira-mono": "^4.5.8"
-  }
+  "type": "module"
 }
