@@ -1,12 +1,7 @@
 <script lang="ts">
   import type { IProject } from "./types"
 
-  export let id: string
-  export let name: string
-  export let image: string
-  export let description = ""
-  export let links: IProject["links"]
-  export let color = "white"
+  const { name, image, description, links, color }: IProject = $props()
 </script>
 
 <div class="rounded-xl overflow-hidden cursor-pointer group w-72 relative transition-color">
@@ -27,7 +22,7 @@
       <p class="text-2xl font-medium">{name}</p>
       <p>{description}</p>
       <div class="flex justify-end mt-1">
-        {#each links as { link, title }}
+        {#each links || [] as { link, title } (link)}
           <a href={link} class="link">{title}</a>
         {/each}
       </div>
@@ -35,10 +30,11 @@
   </div>
 </div>
 
-<style lang="scss">
-  .link {
-    @apply float-right text-blue-600 border border-blue-600 rounded-md px-4 py-1 ml-2
-    transition-colors duration-300 ease-in-out
-    hover:bg-blue-500 hover:text-white;
-  }
+<style lang="postcss">
+    @reference "tailwindcss";
+    .link {
+        @apply float-right text-blue-600 border border-blue-600 rounded-md px-4 py-1 ml-2
+        transition-colors duration-300 ease-in-out
+        hover:bg-blue-500 hover:text-white;
+    }
 </style>
