@@ -1,24 +1,11 @@
-import preprocess from "svelte-preprocess"
-import tailwindcss from "tailwindcss"
-import autoprefixer from "autoprefixer"
 import adapter from "@sveltejs/adapter-static"
+import { vitePreprocess } from "@sveltejs/vite-plugin-svelte"
 
-/** @type {import('@sveltejs/kit').Config} */
+/** @type {import("@sveltejs/kit").Config} */
 const config = {
   // Consult https://github.com/sveltejs/svelte-preprocess
   // for more information about preprocessors
-  preprocess: preprocess({
-    defaults: {
-      style: "scss",
-    },
-    postcss: {
-      plugins: [tailwindcss, autoprefixer],
-    },
-    scss: {
-      prependData: `@import 'src/tailwind.scss';`,
-      outputStyle: "compressed",
-    },
-  }),
+  preprocess: vitePreprocess(),
 
   kit: {
     adapter: adapter({
@@ -26,13 +13,6 @@ const config = {
       assets: "build",
       // fallback: "404.html",
     }),
-    prerender: {
-      default: true,
-      crawl: true,
-      enabled: true,
-      entries: ["*"],
-      onError: "continue",
-    },
   },
 }
 
